+++ app/components/workbench/ExpoQrModal.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Dialog, DialogTitle, DialogDescription, DialogRoot } from '~/components/ui/Dialog';
 import { useStore } from '@nanostores/react';
-import { expoUrlAtom } from '~/stores/qrCodeStore';
+import { expoUrlAtom } from '~/lib/stores/qrCodeStore';
 import QRCode from 'react-qr-code';
 
 interface ExpoQrModalProps {
