+++ app/lib/stores/qrCodeStore.ts
